requirements.txt
requests>=2.31.0
pytest>=8.0.0
python-dotenv>=1.0.1
allure-pytest
